# openapi_spec.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Amplifier App API",
...
        "version": "0.3.0"
    },
    "paths": {
...
        "/applications": {
            "get": {
                "tags": [
                    "applications"
                ],
                "summary": "List all applications",
                "description": "Get a list of all registered applications (without API keys).",
                "operationId": "list_applications_applications_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/ApplicationInfo"
                                    },
                                    "type": "array",
                                    "title": "Response List Applications Applications Get"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "applications"
                ],
                "summary": "Register a new application",
                "description": "Register a new application to access the API.\n\nReturns an API key that must be saved - it won't be shown again!\nThe API key should be sent in the X-API-Key header on all requests.\n\nExample:\n```bash\ncurl -X POST /applications \\\n  -H \"Content-Type: application/json\" \\\n  -d '{\n    \"app_id\": \"mobile-app\",\n    \"app_name\": \"Mobile App\"\n  }'\n```\n\nResponse includes the API key - store it securely!",
                "operationId": "create_application_applications_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ApplicationCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApplicationResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/sessions": {
            "post": {
                "tags": [
                    "sessions"
                ],
                "summary": "Create a new session from a config",
                "description": "Create a new Amplifier session from an existing config.\n\nA session is a lightweight runtime instance that:\n- References a config (complete YAML bundle) via config_id\n- Maintains conversation transcript\n- Tracks execution status\n- Stores runtime state\n\nMultiple sessions can be created from the same config, enabling:\n- Parallel conversations with same configuration\n- A/B testing different approaches\n- Isolated execution contexts\n\nThe first session from a config prepares the bundle (slower).\nSubsequent sessions reuse the cached prepared bundle (faster).\n\nRequirements:\n- The config_id must exist (create via POST /configs first)\n- The config must have valid YAML with required sections\n\nReturns a unique session_id that can be used to:\n- Send messages (POST /sessions/{id}/messages)\n- Stream responses (POST /sessions/{id}/stream)\n- Resume later (POST /sessions/{id}/resume)\n- Delete when done (DELETE /sessions/{id})",
                "operationId": "create_session_sessions_post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SessionCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SessionResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "sessions"
                ],
                "summary": "List Sessions",
                "description": "List all sessions.\n\nArgs:\n    limit: Maximum number of sessions to return\n    offset: Offset for pagination\n\nReturns:\n    SessionListResponse: List of session information",
                "operationId": "list_sessions_sessions_get",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 50,
                            "title": "Limit"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Offset"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SessionListResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/configs": {
            "post": {
                "tags": [
                    "configs"
                ],
                "summary": "Create a new config",
                "description": "Create a new config (complete bundle configuration).\n\nA config contains everything needed to start an Amplifier session.\nConfigs are reusable - create once, use for multiple sessions.\n\n**Shorthand Support:** You can use shorthand names for pre-installed modules:\n- `\"orchestrator\": \"loop-streaming\"` \u00e2\u2020\u2019 uses pre-installed module\n- `\"context\": \"context-simple\"` \u00e2\u2020\u2019 uses pre-installed module\n\nExample:\n```json\n{\n  \"name\": \"my-config\",\n  \"config_data\": {\n    \"bundle\": {\"name\": \"my-config\", \"version\": \"1.0.0\"},\n    \"includes\": [{\"bundle\": \"foundation\"}]\n  },\n  \"session\": {\n    \"orchestrator\": \"loop-streaming\",\n    \"context\": \"context-simple\"\n  },\n  \"providers\": [{\n    \"module\": \"provider-anthropic\",\n    \"config\": {\"api_key\": \"${ANTHROPIC_API_KEY}\", \"model\": \"claude-sonnet-4-5\"}\n  }]\n}\n```",
                "operationId": "create_config_configs_post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ConfigCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ConfigResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "configs"
                ],
                "summary": "List Configs",
                "description": "List all configs.\n\nArgs:\n    limit: Maximum number of configs to return\n    offset: Offset for pagination\n\nReturns:\n    ConfigListResponse: List of config metadata",
                "operationId": "list_configs_configs_get",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 50,
                            "title": "Limit"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Offset"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ConfigListResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/smoke-tests": {
            "get": {
                "tags": [
                    "testing"
                ],
                "summary": "Run Smoke Tests",
                "description": "Run smoke tests and return results.\n\nThis endpoint executes pytest on the smoke test suite and returns results\nin JSON format. Useful for automated health checks and CI/CD pipelines.\n\nArgs:\n    verbose: If true, include detailed test output\n    pattern: Test file pattern (default: test_smoke.py)\n\nReturns:\n    JSON with test results including pass/fail counts and details",
                "operationId": "run_smoke_tests_smoke_tests_get",
                "parameters": [
                    {
                        "name": "verbose",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "description": "Show verbose test output",
                            "default": false,
                            "title": "Verbose"
                        },
                        "description": "Show verbose test output"
                    },
                    {
                        "name": "pattern",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "Test file pattern to run",
                            "default": "test_smoke.py",
                            "title": "Pattern"
                        },
                        "description": "Test file pattern to run"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": true,
                                    "title": "Response Run Smoke Tests Smoke Tests Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApplicationCreate": {
                "properties": {
                    "app_id": {
                        "type": "string",
                        "pattern": "^[a-z0-9-]+$",
                        "title": "App Id",
                        "description": "Unique application identifier",
                        "examples": [
                            "mobile-app",
                            "web-app",
                            "desktop-app"
                        ]
                    },
                    "app_name": {
                        "type": "string",
                        "title": "App Name",
                        "description": "Human-readable name",
                        "examples": [
                            "Mobile App",
                            "Web App"
                        ]
                    },
                    "settings": {
                        "additionalProperties": true,
                        "type": "object",
                        "title": "Settings",
                        "description": "Optional application settings"
                    }
                },
                "type": "object",
                "required": [
                    "app_id",
                    "app_name"
                ],
                "title": "ApplicationCreate",
                "description": "Request to create a new application."
            },
            "ApplicationInfo": {
                "properties": {
                    "app_id": {
                        "type": "string",
                        "title": "App Id"
                    },
                    "app_name": {
                        "type": "string",
                        "title": "App Name"
                    },
                    "is_active": {
                        "type": "boolean",
                        "title": "Is Active"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Updated At"
                    }
                },
                "type": "object",
                "required": [
                    "app_id",
                    "app_name",
                    "is_active",
                    "created_at",
                    "updated_at"
                ],
                "title": "ApplicationInfo",
                "description": "Application information (without sensitive data)."
            },
            "ApplicationResponse": {
                "properties": {
                    "app_id": {
                        "type": "string",
                        "title": "App Id"
                    },
                    "app_name": {
                        "type": "string",
                        "title": "App Name"
                    },
                    "api_key": {
                        "type": "string",
                        "title": "Api Key",
                        "description": "API key - save this! Won't be shown again"
                    },
                    "is_active": {
                        "type": "boolean",
                        "title": "Is Active"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "default": "Application registered successfully"
                    }
                },
                "type": "object",
                "required": [
                    "app_id",
                    "app_name",
                    "api_key",
                    "is_active",
                    "created_at"
                ],
                "title": "ApplicationResponse",
                "description": "Response when creating an application."
            },
...
            "ConfigListResponse": {
                "properties": {
                    "configs": {
                        "items": {
                            "$ref": "#/components/schemas/ConfigMetadata"
                        },
                        "type": "array",
                        "title": "Configs"
                    },
                    "total": {
                        "type": "integer",
                        "title": "Total"
                    }
                },
                "type": "object",
                "required": [
                    "configs",
                    "total"
                ],
                "title": "ConfigListResponse",
                "description": "Response with list of configs."
            },
...
            "ConfigResponse": {
                "properties": {
                    "config_id": {
                        "type": "string",
                        "title": "Config Id"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description"
                    },
                    "config_data": {
                        "additionalProperties": true,
                        "type": "object",
                        "title": "Config Data"
                    },
                    "user_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "User Id"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Updated At"
                    },
                    "tags": {
                        "additionalProperties": {
                            "type": "string"
                        },
                        "type": "object",
                        "title": "Tags"
                    },
                    "message": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Message"
                    }
                },
                "type": "object",
                "required": [
                    "config_id",
                    "name",
                    "config_data",
                    "created_at",
                    "updated_at"
                ],
                "title": "ConfigResponse",
                "description": "Response with config data."
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "SessionCreateRequest": {
                "properties": {
                    "config_id": {
                        "type": "string",
                        "title": "Config Id",
                        "description": "Config ID to use for this session (must exist)",
                        "examples": [
                            "c7a3f9e2-1b4d-4c8a-9f2e-d6b8a1c5e3f7"
                        ]
                    },
                    "tags": {
                        "additionalProperties": {
                            "type": "string"
                        },
                        "type": "object",
                        "title": "Tags",
                        "description": "Optional metadata tags for this session",
                        "examples": [
                            {
                                "project": "my-app",
                                "user": "john"
                            }
                        ]
                    }
                },
                "type": "object",
                "required": [
                    "config_id"
                ],
                "title": "SessionCreateRequest",
                "description": "Request to create a new session from a config.\n\nA session is a runtime instance that references a config (complete YAML bundle).\nMultiple sessions can be created from the same config."
            },
...
            "SessionListResponse": {
                "properties": {
                    "sessions": {
                        "items": {
                            "$ref": "#/components/schemas/SessionInfo"
                        },
                        "type": "array",
                        "title": "Sessions"
                    },
                    "total": {
                        "type": "integer",
                        "title": "Total"
                    }
                },
                "type": "object",
                "required": [
                    "sessions",
                    "total"
                ],
                "title": "SessionListResponse",
                "description": "Response for listing sessions."
            },
            "SessionResponse": {
                "properties": {
                    "session_id": {
                        "type": "string",
                        "title": "Session Id"
                    },
                    "config_id": {
                        "type": "string",
                        "title": "Config Id"
                    },
                    "status": {
                        "$ref": "#/components/schemas/SessionStatus"
                    },
                    "message": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Message"
                    }
                },
                "type": "object",
                "required": [
                    "session_id",
                    "config_id",
                    "status"
                ],
                "title": "SessionResponse",
                "description": "Response for session operations."
            },
            "SessionStatus": {
                "type": "string",
                "enum": [
                    "active",
                    "completed",
                    "failed",
                    "cancelled"
                ],
                "title": "SessionStatus",
                "description": "Session status enumeration."
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        }
    }
}